+++ C#/ACE/AES.cs
@@ -38,9 +38,8 @@ namespace ACE
         public static AESHelper AESHelper = new AESHelper(fileRegisterinfo);
         public static string? GetSysConfig()
         {
-            string path = "application.xml";
-
-            return File.ReadAllText(path)??null;
+            //string path = "application.xml";
+            //return File.ReadAllText(path)??null;
             if (!IsInit) Init();
             byte[] trest1 = AESHelper.Decrypt();
             if (trest1 == null)
@@ -56,7 +55,7 @@ namespace ACE
 
         public static string? GetExpireDate()
         {
-            return "2025/1/1";
+            //return "2025/1/1";
             if (!IsInit) Init();
 
             if (!File.Exists("application.xml"))