+++ C#/Solution/OpenSolutionWindow.xaml.cs
@@ -20,10 +20,9 @@ namespace Solution
         }
         public string FullName = string.Empty;
 
-        RecentFileList recentFileList = new RecentFileList();
+        RecentFileList recentFileList = new RecentFileList() { Persister = new RegistryPersister("Software\\NLG\\Grid\\SolutionHistory") };
+
         public ObservableCollection<SoulutionInfo> SoulutionInfos = new ObservableCollection<SoulutionInfo>();
-
-
         private void BaseWindow_Initialized(object sender, EventArgs e)
         {
             foreach (var item in recentFileList.RecentFiles)